server working, dft/read refactored to stream rather than batch

+++ dft.py
@@ -14,9 +14,10 @@ conn = sqlite3.connect('data.db')
 def get_message(fifo: int) -> str:
     frame = os.read(fifo, 2 * N)
     msg = [ x for x in frame ]#[ int.from_bytes(x, byteorder='little') for x in msg ]
+    msg = [(3.3 * x / 4095) for x in msg ]
     data = np.array([])
-    for i in range(0, N * 2, 2):
-        data = np.append(data, msg[i] + msg[i + 1] * 16**2)
+    for i in range(0, len(msg), 2):
+        data = np.append(data, msg[i + 1] + msg[i] * 16**2)
     return data
 
 def dft(batch):
@@ -50,18 +51,26 @@ except Exception as e:
     exit()
 
 try:
+    batch = np.array([])
+    i = 0
     while True:
         if fifo:
-            conn.execute('delete from samples')
-            batch = get_message(fifo)
-            for i,x in enumerate(batch):
-                conn.execute('insert into samples (sample, value) VALUES (%d, %f)' % (i,x))
-            conn.commit()
-            real, imag = dft(batch)
-            conn.execute('delete from dft')
-            for i, (x,y) in enumerate(zip(real, imag)):
-                conn.execute('insert into dft (bucket, real, imag) VALUES (%d, %f, %f)' % (i,x,y))
-            conn.commit()
+            new_batch = get_message(fifo)
+            i += len(new_batch)
+            batch = np.append(batch, new_batch)
+            if i >= N:
+                batch = batch[0:64]
+                conn.execute('delete from samples')
+                for i,x in enumerate(batch):
+                    conn.execute('insert into samples (sample, value) VALUES (%d, %f)' % (i,x))
+                conn.commit()
+                real, imag = dft(batch)
+                conn.execute('delete from dft')
+                for i, (x,y) in enumerate(zip(real, imag)):
+                    conn.execute('insert into dft (bucket, real, imag) VALUES (%d, %f, %f)' % (i,x,y))
+                conn.commit()
+                batch = []
+                i = 0
 except Exception as e:
     print(traceback.format_exc())
 finally:
